TestSulWowza: refactor verifyStacksToken tests into separate file
